theatrical release part ok
