+++ libs/movie/src/lib/movie/form/sales-info/theatrical-release/theatrical-release.component.html
@@ -1,0 +1,15 @@
+<form [formGroup]="form">
+  <article fxLayout="row" fxLayoutAlign="space-between center">
+    <p>Theatrical Release</p>
+    <mat-slide-toggle color="primary" formControlName="theatricalRelease"></mat-slide-toggle>
+  </article>
+
+  <article fxLayout="column">
+    <mat-form-field appearance="outline">
+      <mat-label>Release Date in Country of Origin</mat-label>
+      <input matInput formControlName="originCountryReleaseDate" [matDatepicker]="originReleaseDate" placeholder="MM/DD/YYYY">
+      <mat-datepicker-toggle matSuffix [for]="originReleaseDate"></mat-datepicker-toggle>
+      <mat-datepicker #originReleaseDate></mat-datepicker>
+    </mat-form-field>
+  </article>
+</form>
