--- libs/movie/src/lib/movie/form/theatrical-release/theatrical-release.component.html
@@ -1,13 +1,0 @@
-<article fxLayout="row" fxLayoutAlign="space-between center">
-  <p>Theatrical Release</p>
-  <mat-slide-toggle color="primary" formControlName="theatricalRelease"></mat-slide-toggle>
-</article>
-
-<article fxLayout="column">
-  <mat-form-field appearance="outline">
-    <mat-label>Release Date in Country of Origin</mat-label>
-    <input matInput [formControl]="originCountryReleaseDate" [matDatepicker]="originReleaseDate" placeholder="MM/DD/YYYY">
-    <mat-datepicker-toggle matSuffix [for]="originReleaseDate"></mat-datepicker-toggle>
-    <mat-datepicker #originReleaseDate></mat-datepicker>
-  </mat-form-field>
-</article>
